+++ both/lib/Config.jsx
@@ -86,6 +86,10 @@ var navFM = [{
 	path:"requests",		
 	label:"Requests",
 	icon:"fa fa-wrench"
+},{
+	path:"reports",		
+	label:"Reports",
+	icon:"fa fa-bar-chart"
 }];
 
 if(false) {