+++ packages/fmc-reports/routes.jsx
@@ -1,0 +1,34 @@
+import React from "react";
+import {mount} from 'react-mounter';
+
+loggedIn.route('/reports', {
+  name: 'reports',
+  action(params) {
+    mount(MainLayout,{content:<ReportsPageIndex/>});
+  }
+});
+
+loggedIn.route('/report/:reportId', {
+  name: 'report',
+  action(params) {
+    mount(MainLayout,{content:<ReportsPageSingle id={params.reportId}/>});
+  }
+});
+
+loggedIn.route('/report/:reportId/:view', {
+  name: 'report',
+  action(params) {
+    var Layout = MainLayout;
+    if(params.view=="print") {
+      Layout = PrintLayout;
+    }
+    mount(Layout,{content:<ReportsPageSingle id={params.reportId}/>});
+  }
+});
+
+loggedIn.route('/dashboard', {
+  name: 'dashboard',
+  action() {
+    mount(MainLayout,{content:<DashboardPage/>});
+  }
+});